+++ slurmweb/errors.py
@@ -35,3 +35,19 @@ class SlurmwebRestdError(Exception):
 
     def __str__(self):
         return f"SlurwebRestdError({self.message}, {self.error}, {self.description}, {self.source})"
+
+
+# Alias JSONDecodeError from simplejson external library and json standard library
+# module to catch generically the error raised with Requests < 2.27 on old systems in
+# presence of unexepected non-JSON responses.
+#
+# This is not needed with Requests >= 2.27 where the same logic is implemented with
+# requests.exceptions.JSONDecodeError wildcard exception. For reference, see:
+# https://github.com/psf/requests/pull/5856
+
+try:
+    from simplejson import JSONDecodeError
+except ImportError:
+    from json import JSONDecodeError
+
+SlurmwebCompatJSONDecodeError = JSONDecodeError